services/agent/requirements.txt
livekit-agents[google]~=1.3
livekit-plugins-google~=1.3
supabase~=2.15
python-dotenv~=1.0
aiohttp~=3.11

